requirements.txt
icalendar==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
tzdata==2024.2
